backend/src/auth_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
google-auth==2.29.0
requests==2.32.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
asyncpg==0.29.0
pydantic-settings==2.3.3
